Fixed a bug with the component filters not working as expected
A customer reported a bug where the component filters were not being properly applied to subcomponents.

If a customer listed subcomponents in the `data-filter-components` attribute notices would be displayed for other sibling components.

This was due to an incorrect and overly complex query being made to the API that was querting for notices elsewhere in the components family tree.

Historically this worked as expected, but we fixed another bug in the API itself last year which appears to have create some inconsistent behaviour in the plugin.

This should now behave as intended.